+++ lib/managers/window_manager.dart
@@ -12,7 +12,7 @@ class WindowManager = WindowManagerBase with _$WindowManager;
 abstract class WindowManagerBase extends Subsystem with Store, Logger {
 
   @override
-  String subsystemName = "Window Manager";
+  String subsystemName = "Window manager";
 
   @readonly
   bool _isFullScreen = false;